+++ examples/RPC/client.php
@@ -8,8 +8,6 @@ use Rubix\Server\Commands\Proba;
 use Rubix\Server\Commands\ProbaSample;
 use Rubix\Server\Commands\Predict;
 use Rubix\Server\Commands\PredictSample;
-use Rubix\Server\Commands\QueryModel;
-use Rubix\Server\Commands\ServerStatus;
 
 $client = new RPCClient('127.0.0.1', 8888, false, [
     'Authorization' => 'Basic ' . base64_encode('test:secret'),
@@ -25,5 +23,3 @@ print_r($client->send(new Predict($dataset)));
 print_r($client->send(new PredictSample($dataset->sample(0))));
 print_r($client->send(new Proba($dataset)));
 print_r($client->send(new ProbaSample($dataset->sample(0))));
-print_r($client->send(new QueryModel()));
-print_r($client->send(new ServerStatus()));