+++ tests/RESTServerTest.php
@@ -36,10 +36,4 @@ class RESTServerTest extends TestCase
         $this->assertInstanceOf(RESTServer::class, $this->server);
         $this->assertInstanceOf(Server::class, $this->server);
     }
-
-    protected function assertPreConditions() : void
-    {
-        $this->assertEquals(0, $this->server->requests());
-        $this->assertEquals(0, $this->server->uptime());
-    }
 }